Moves test_distribute_cells_multi_pop_multi_cycle to unit tests
