--- tests/integration-e2e/test_dry_run_workflow.py
@@ -1,243 +1,0 @@
-import json
-import platform
-import pytest
-import tempfile
-from pathlib import Path
-from unittest.mock import patch
-
-from tests.utils import defaultdict_to_standard_types
-from neurodamus.utils.memory import DryRunStats
-from neurodamus.core.configuration import GlobalConfig, LogLevel
-
-SIM_DIR = Path(__file__).parent.parent.absolute() / "simulations"
-TMP_FOLDER = tempfile.mkdtemp()
-
-
-@pytest.fixture(autouse=True)
-def change_test_dir(monkeypatch):
-    """
-    All tests in this file are using the same working directory, i.e TMP_FOLDER
-    Because test_dynamic_distribute requires memory_per_metype.json generated in the previous test
-    """
-    monkeypatch.chdir(TMP_FOLDER)
-
-
-@pytest.fixture
-def neurodamus_instance(request: pytest.FixtureRequest, USECASE3: Path):
-    from neurodamus import Neurodamus
-
-    params = request.param
-    dry_run = params.get('dry_run', True)
-    num_target_ranks = params.get('num_target_ranks', '1')
-    modelbuilding_steps = params.get('modelbuilding_steps', '1')
-    config_file = Path(params.get('config_file', "simulation_sonata.json")).name
-    path_to_config = params.get('path_to_config', USECASE3)
-    lb_mode = params.get('lb_mode', "")
-
-    with open(path_to_config / config_file) as src_f:
-        sim_config_data = json.load(src_f)
-    sim_config_data["network"] = str(path_to_config / "circuit_config.json")
-    sim_config_data["node_sets_file"] = str(path_to_config / "nodesets.json")
-    with open(config_file, "w") as dst_f:
-        json.dump(sim_config_data, dst_f, indent=2)
-
-    GlobalConfig.verbosity = LogLevel.DEBUG
-    nd = Neurodamus(
-        str(config_file),
-        dry_run=dry_run,
-        num_target_ranks=num_target_ranks,
-        modelbuilding_steps=modelbuilding_steps,
-        lb_mode=lb_mode
-    )
-    yield nd
-
-    nd = None
-
-
-@pytest.mark.parametrize("neurodamus_instance", [
-    {
-        'dry_run': True,
-        'num_target_ranks': 2,
-        'config_file': "simulation_sonata.json",
-        'path_to_config': SIM_DIR / "usecase3",
-        'lb_mode': ""
-    }
-], indirect=True)
-def test_dry_run_workflow(neurodamus_instance):
-    """
-    Test that the dry run mode works
-
-    """
-    isMacOS = platform.system() == "Darwin"
-    from neurodamus.utils.memory import export_allocation_stats
-    from neurodamus.utils.memory import export_metype_memory_usage
-
-    GlobalConfig.verbosity = LogLevel.DEBUG
-    nd = neurodamus_instance
-
-    nd.run()
-
-    # the following uncommented line should be investigated with:
-    # https://github.com/openbraininstitute/neurodamus/issues/44?issue=openbraininstitute%7Cneurodamus%7C45
-    # it should be max 30 but on mac it consumes a little more. Also, sometimes,
-    # the CI fails for this. For now I increase the limit to 40 as a workaround.
-    assert 20.0 <= nd._dry_run_stats.cell_memory_total <= 40
-    assert 0.0 <= nd._dry_run_stats.synapse_memory_total <= 1.0
-    # the following uncommented line should be investigated with:
-    # https://github.com/openbraininstitute/neurodamus/issues/44?issue=openbraininstitute%7Cneurodamus%7C45
-    # it should be min 55 but on mac it consumes a little more. For now I decrease the limit to
-    # to 50 as a workaround
-    assert (50.0 if isMacOS else 55) <= nd._dry_run_stats.base_memory <= 120.0
-    expected_items = {
-        'L4_PC-dSTUT': 2,
-        'L4_MC-dSTUT': 1,
-        'L4_MC-dNAC': 1,
-        'L5_PC-dSTUT': 1
-    }
-    assert nd._dry_run_stats.metype_counts == expected_items
-    assert nd._dry_run_stats.suggest_nodes(0.3) > 0
-
-    # Test that the allocation works and can be saved and loaded
-    rank_alloc, _, cell_mem_use = nd._dry_run_stats.distribute_cells_with_validation(2, 1, None)
-    export_allocation_stats(rank_alloc,
-                            "allocation", 2, 1)
-    export_metype_memory_usage(cell_mem_use, "memory_per_metype.json")
-
-    rank_alloc = nd._dry_run_stats.import_allocation_stats("allocation_r2_c1.pkl.gz", 0)
-    rank_allocation_standard = defaultdict_to_standard_types(rank_alloc)
-
-    expected_items = {
-        'NodeA': {(0, 0): [1]},
-        'NodeB': {(0, 0): [1]}
-    }
-
-    assert rank_allocation_standard == expected_items
-
-    # Test that the allocation works and can be saved and loaded
-    # and generate allocation file for 1 rank
-    rank_alloc, _, cell_mem_use = nd._dry_run_stats.distribute_cells_with_validation(1, 1, None)
-    export_metype_memory_usage(cell_mem_use, "memory_per_metype.json")
-    rank_allocation_standard = defaultdict_to_standard_types(rank_alloc)
-
-    expected_items = {
-        'NodeA': {(0, 0): [1, 2, 3]},
-        'NodeB': {(0, 0): [1, 2]}
-    }
-
-    assert rank_allocation_standard == expected_items
-
-    Path(("allocation_r1_c1.pkl.gz")).unlink(missing_ok=True)
-    Path(("allocation_r2_c1.pkl.gz")).unlink(missing_ok=True)
-
-
-@pytest.mark.parametrize("neurodamus_instance", [
-    {
-        'dry_run': False,
-        'config_file': "simulation_sonata.json",
-        'path_to_config': SIM_DIR / "usecase3",
-        'lb_mode': "Memory"
-    }
-], indirect=True)
-def test_dynamic_distribute(neurodamus_instance):
-    """
-    Test that the dynamic distribution of cells works properly.
-    The test deletes any old allocation file before running and uses
-    the memory_per_metype.json generated in the previous test to
-    redistribute the cells. Then checks if the new allocation is correct.
-    """
-    nd = neurodamus_instance
-
-    rank_allocation, _, _ = nd._dry_run_stats.distribute_cells_with_validation(2, 1)
-    rank_allocation_standard = defaultdict_to_standard_types(rank_allocation)
-
-    expected_items = {
-        'NodeA': {
-            (0, 0): [1],
-            (1, 0): [2, 3]
-        }
-    }
-
-    for key, sub_value in rank_allocation_standard['NodeA'].items():
-        assert set(sub_value) == set(expected_items['NodeA'][key])
-
-
-@pytest.fixture
-def fixed_memory_measurements():
-    with patch('neurodamus.utils.memory.get_mem_usage_kb', return_value=100):
-        with patch('neurodamus.utils.memory.SynapseMemoryUsage.get_memory_usage', return_value=10):
-            yield
-
-
-def test_distribute_cells_multi_pop_multi_cycle(fixed_memory_measurements):
-    """
-    Test that the distribute_cells_with_validation function works with multiple pops and cycles
-    """
-    stats = DryRunStats()
-
-    # Mock data for testing
-    stats.metype_memory = {
-        'L4_PC-dSTUT': 50,
-        'L4_MC-dSTUT': 30,
-        'L4_MC-dNAC': 20,
-        'L5_PC-dSTUT': 40
-    }
-    stats.metype_cell_syn_average = {
-        'L4_PC-dSTUT': 5,
-        'L4_MC-dSTUT': 3,
-        'L4_MC-dNAC': 2,
-        'L5_PC-dSTUT': 4
-    }
-    stats.pop_metype_gids = {
-        'NodeA': {
-            'L4_PC-dSTUT': [1, 2, 3],
-            'L4_MC-dSTUT': [4, 5],
-            'L4_MC-dNAC': [6],
-            'L5_PC-dSTUT': [7, 8]
-        },
-        'NodeB': {
-            'L4_PC-dSTUT': [9],
-            'L4_MC-dSTUT': [10, 11],
-            'L4_MC-dNAC': [12, 13],
-            'L5_PC-dSTUT': [14]
-        }
-    }
-
-    # Run the distribute_cells_with_validation function
-    bucket_allocation, bucket_memory, metype_memory_usage = stats.distribute_cells_with_validation(
-        num_ranks=2,
-        cycles=2
-    )
-    rank_allocation_standard = defaultdict_to_standard_types(bucket_allocation)
-
-    expected_allocation = {
-        'NodeA': {
-            (0, 0): [1, 6],
-            (1, 0): [3, 8],
-            (0, 1): [2, 7],
-            (1, 1): [4, 5]
-        },
-        'NodeB': {
-            (0, 0): [9],
-            (1, 0): [11],
-            (0, 1): [10, 14],
-            (1, 1): [12, 13]
-        }
-    }
-    expected_memory = {
-        'NodeA': {
-            (0, 0): 90,
-            (1, 0): 110,
-            (0, 1): 110,
-            (1, 1): 80
-        },
-        'NodeB': {
-            (0, 0): 60,
-            (1, 0): 40,
-            (0, 1): 90,
-            (1, 1): 60
-        }
-    }
-
-    # Assert that the results match the expected values
-    assert rank_allocation_standard == expected_allocation
-    assert bucket_memory == expected_memory
